Moved main method for table game
